+++ src/lib/input-validation/constant/numbers.ts
@@ -4,8 +4,7 @@
  * and restrictions contact your company contract manager.
  */
 
-export const MAX_SAFE_INTEGER = 9007199254740991;
-export const JAVA_MAX_INT = 2147483647;
+export const MAX_INTEGER = 2147483647;
 export const MAX_LONG_TEXT_LENGTH = 2000;
 export const MAX_SHORT_TEXT_LENGTH = 256;
 export const MAX_DISPLAY_NAME_LENGTH = 48;